eventos/requirements.txt
fastapi
uvicorn[standard]
kafka-python  # Para el consumidor Kafka
pymongo       # Driver de Python para MongoDB
python-dotenv # Para cargar variables de entorno
